feat: move write blocks for docker-compose
