+++ generators/docker-compose/index.js
@@ -8,6 +8,7 @@ var generators = require('yeoman-generator'),
     pathjs = require('path'),
     util = require('util'),
     prompts = require('./prompts'),
+    writeFiles = require('./files').writeFiles,
     scriptBase = require('../generator-base');
 
 var DockerComposeGenerator = generators.Base.extend({});
@@ -309,39 +310,8 @@ module.exports = DockerComposeGenerator.extend({
         }
     },
 
-    writing: {
-        writeDockerCompose: function() {
-            this.template('_docker-compose.yml', 'docker-compose.yml');
-        },
+    writing: writeFiles(),
 
-        writeRegistryFiles: function() {
-            if(this.gatewayNb === 0 && this.microserviceNb === 0) return;
-            if(this.serviceDiscoveryType === 'eureka'){
-                this.template('_jhipster-registry.yml', 'jhipster-registry.yml');
-            }
-            if(this.serviceDiscoveryType === 'consul'){
-                this.template('_consul.yml', 'consul.yml');
-                this.copy('consul-conf/_acl_config.json', 'consul-conf/acl_config.json');
-            }
-            if(this.serviceDiscoveryType){
-                this.template('central-server-config/_application.yml', 'central-server-config/application.yml');
-            }
-        },
-
-        writeKafkaFiles: function() {
-            if(!this.useKafka) return;
-
-            this.template('_kafka.yml', 'kafka.yml');
-        },
-
-        writeElkFiles: function() {
-            if(!this.useElk) return;
-
-            this.copy('_jhipster-console.yml', 'jhipster-console.yml');
-            this.copy('log-conf/_logstash.conf', 'log-conf/logstash.conf');
-            this.copy('log-data/_.gitignore', 'log-data/.gitignore');
-        }
-    },
     end: function() {
         if (this.warning) {
             this.log('\n');
